Move brick volfile template to volgen package
